+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+
+- Fix: Remove `isRequired` of the `ownerName` props on `Comment` component.
 - Fix: Update
   `Comment`,
   `CommentForm`,